fix: Invoke Callback steps
